requirements.txt
asgiref==3.8.1
cffi==1.16.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
